+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 2.0.0 (Next)
+### 2.0.0
 **General**
 - Global variable methods and associated parameters converted to camelCase.
 - Remove support for old DSL/closure type inputs to global variable methods.